# dfirtrack_api/openapi/openapi_dfirtrack.yml
openapi: 3.0.2
info:
  title: DFIRTrack
  description: OpenAPI 3 - Documentation of DFIRTrack API
  version: v1.2.1
paths:
...
  /api/case/:
    get:
      operationId: listCases
      description: 'all objects, allowed: GET + POST'
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Case'
          description: ''
      tags:
      - api
    post:
      operationId: createCase
      description: 'all objects, allowed: GET + POST'
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Case'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Case'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Case'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
          description: ''
      tags:
      - api
...
  /api/system/:
    get:
      operationId: listSystems
      description: 'all objects, allowed: GET + POST'
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/System'
          description: ''
      tags:
      - api
    post:
      operationId: createSystem
      description: 'all objects, allowed: GET + POST'
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/System'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/System'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/System'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/System'
          description: ''
      tags:
      - api
  /api/system/{system_id}/:
    get:
      operationId: retrieveSystem
      description: 'single object, allowed: GET + PUT'
      parameters:
      - name: system_id
        in: path
        required: true
        description: A unique integer value identifying this system.
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/System'
          description: ''
      tags:
      - api
    put:
      operationId: updateSystem
      description: 'single object, allowed: GET + PUT'
      parameters:
      - name: system_id
        in: path
        required: true
        description: A unique integer value identifying this system.
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/System'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/System'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/System'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/System'
          description: ''
      tags:
      - api
    patch:
      operationId: partialUpdateSystem
      description: 'single object, allowed: GET + PUT'
      parameters:
      - name: system_id
        in: path
        required: true
        description: A unique integer value identifying this system.
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/System'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/System'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/System'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/System'
          description: ''
      tags:
      - api
...
  /api/tag/:
    get:
      operationId: listTags
      description: 'all objects, allowed: GET + POST'
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
          description: ''
      tags:
      - api
    post:
      operationId: createTag
      description: 'all objects, allowed: GET + POST'
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Tag'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
          description: ''
      tags:
      - api
...
components:
  schemas:
...
    Case:
      type: object
      properties:
        case_id:
          type: integer
          readOnly: true
        case_id_external:
          type: string
          nullable: true
          maxLength: 50
        case_name:
          type: string
          maxLength: 50
        casepriority:
          type: integer
        casestatus:
          type: integer
        casetype:
          type: integer
          nullable: true
        tag:
          type: array
          items:
            type: integer
        case_is_incident:
          type: boolean
        case_start_time:
          type: string
          format: date-time
          nullable: true
        case_end_time:
          type: string
          format: date-time
          nullable: true
        case_created_by_user_id:
          type: integer
        case_create_time:
          type: string
          format: date-time
          readOnly: true
        case_modified_by_user_id:
          type: integer
          nullable: true
        case_modify_time:
          type: string
          format: date-time
          readOnly: true
      required:
      - case_name
      - case_is_incident
      - case_created_by_user_id
...
    System:
      type: object
      properties:
        system_id:
          type: integer
          readOnly: true
        system_uuid:
          type: string
          format: uuid
          readOnly: true
        system_name:
          type: string
          maxLength: 50
        domain:
          type: integer
          nullable: true
        dnsname:
          type: integer
          nullable: true
        systemstatus:
          type: integer
        analysisstatus:
          type: integer
          nullable: true
        reason:
          type: integer
          nullable: true
        recommendation:
          type: integer
          nullable: true
        systemtype:
          type: integer
          nullable: true
        ip:
          type: array
          items:
            type: integer
        os:
          type: integer
          nullable: true
        osarch:
          type: integer
          nullable: true
        system_lastbooted_time:
          type: string
          format: date-time
          nullable: true
        system_deprecated_time:
          type: string
          format: date-time
          nullable: true
        system_is_vm:
          type: boolean
          nullable: true
        host_system:
          type: integer
          nullable: true
        company:
          type: array
          items:
            type: integer
        location:
          type: integer
          nullable: true
        serviceprovider:
          type: integer
          nullable: true
        contact:
          type: integer
          nullable: true
        tag:
          type: array
          items:
            type: integer
        case:
          type: array
          items:
            type: integer
        system_create_time:
          type: string
          format: date-time
          readOnly: true
        system_created_by_user_id:
          type: integer
        system_modify_time:
          type: string
          format: date-time
          readOnly: true
        system_modified_by_user_id:
          type: integer
        system_export_markdown:
          type: boolean
        system_export_spreadsheet:
          type: boolean
      required:
      - system_name
      - systemstatus
      - system_created_by_user_id
      - system_modified_by_user_id
...
    Tag:
      type: object
      properties:
        tag_id:
          type: integer
          readOnly: true
        tag_name:
          type: string
          maxLength: 50
        tagcolor:
          type: integer
      required:
      - tag_name
      - tagcolor
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: basic authentication
    bearerAuth:
      type: http
      scheme: bearer
      name: Token
      description: API Token authentication
security:
- basicAuth: []
- bearerAuth: []